\name{p2.2}
\alias{p2.2}
\docType{data}
\title{Problem 2.2 of \dQuote{The New Weibull Handbook, Fifth edition}}
\description{Dataset of five failed parts of eight parts in service.}
\usage{p2.2}
\format{An unsorted dataframe with eight rows and colums \code{$serial} (serial number of the part), \code{$time} ((life)time observation in hours) and \code{$event} (censoring indicator)}
\source{
Robert B. Abernethy,  \emph{The New Weibull Handbook, Fifth Edition} (Robert B. Abernethy, 2004)
}

%\references{
%Robert B. Abernethy,  \emph{The New Weibull Handbook, Fifth Edition} (Robert B. Abernethy, 2004)
%}
\keyword{datasets}