\name{abrem-package}
\alias{abrem-package}
\alias{abrem}
\docType{package}
\title{
Abernethy Reliability Methods
}
\description{
This package will calculate fits from (life-)time observations, following the methods of \emph{The New Weibull Handbook}, Fifth Edition by Dr. Robert B. Abernethy. It is based on the now outdated and unmaintained \pkg{weibulltoolkit} package, still available from \url{http://sourceforge.net/projects/Weibulltoolkit}.
}

\author{
Jurgen Symynck <jurgen.symynck@kahosl.be>

Maintainer: Jurgen Symynck <jurgen.symynck@kahosl.be>
}

\note{}
\section{To Do}{}

\references{
Robert B. Abernethy,  \emph{The New Weibull Handbook, Fifth Edition} (Robert B. Abernethy, 2004)

Jerald F. Lawless,  \emph{Statistical Models and Methods for Lifetime Data, 2nd edition} (Wiley-Interscience, Hoboken  N.J., 2003)

William Q. Meeker and Luis A. Escobar,  \emph{Statistical Methods for Reliability Data} (Wiley-Interscience, New York, 1998)

Chi-Chao Lui, \emph{A Comparison Between The Weibull And Lognormal Models Used To Analyse Reliability Data} (dissertation from University of Nottingham, 1997)

Efron, B. and Tibshirani, R.,  \emph{An Introduction to the Bootstrap} (Chapman & Hall, 1993)

Jurgen Symynck, Filip De Bal,  \emph{Weibull analysis using R, in a nutshell} (New Technologies and Products in Machine Manufacturing Technology, Stefan cel Mare University of Suceava, 2010). \cr Available on \url{http://mechanics.kahosl.be/fatimat/index.php/downloads-and-information/40/171}

Jurgen Symynck, Filip De Bal,  \emph{Monte Carlo pivotal confidence bounds for Weibull analysis, with implementations in R} (New Technologies and Products in Machine Manufacturing Technology, Stefan cel Mare University of Suceava, 2011).\cr Available on \url{http://www.fim.usv.ro/conf_1/tehnomusjournal/pagini/journal2011/files/7.pdf} and \url{http://mechanics.kahosl.be/fatimat/index.php/downloads-and-information/40/198}


\emph{SuperSMITH(TM) Weibull} homepage (\url{http://www.Weibullnews.com/contents.htm})

\url{http://mechanics.kahosl.be/fatimat}

\url{http://r-forge.r-project.org/projects/abernethy/}

\url{http://sourceforge.net/projects/weibulltoolkit/}}

\keyword{ package }


