\name{p2.2}
\alias{p2.2}
\docType{data}
\title{
    Problem 2.2 of \dQuote{The New Weibull Handbook}, Fifth Edition
}
\description{
    Dataset of five failed parts from a total of eight parts in service.
}
\usage{p2.2}
\format{
    An unsorted dataframe with 8 rows, with columns \code{$serial}
    (serial number of the part), \code{$time} ((life-)time observations
    in hours) and \code{$event} (censoring indicator).
}
\source{
    Robert B. Abernethy, \emph{The New Weibull Handbook}, Fifth Edition
    (Robert B. Abernethy, 2004)
}
%\references{
%Robert B. Abernethy,  \emph{The New Weibull Handbook, Fifth Edition} (Robert B. Abernethy, 2004)
%}
\keyword{datasets}